requirements.txt
aniso8601==8.0.0
appnope==0.1.0
arrow==0.15.5
astroid==2.3.3
attrs==19.3.0
aws-lambda-builders==0.7.0
aws-sam-cli==0.43.0
aws-sam-translator==1.20.1
backcall==0.1.0
binaryornot==0.4.4
bleach==3.1.1
boto3==1.12.12
botocore==1.15.12
certifi==2019.11.28
chardet==3.0.4
chevron==0.13.1
click==7.1.1
cookiecutter==1.6.0
dateparser==0.7.2
decorator==4.4.2
defusedxml==0.6.0
dnspython==1.16.0
docker==4.2.0
docutils==0.15.2
entrypoints==0.3
Flask==1.1.2
Flask-Cors==3.0.8
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.0
future==0.18.2
idna==2.8
importlib-metadata==1.6.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.2
jinja2-time==0.2.0
jmespath==0.9.5
jsonify==0.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
logger==1.4
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mysql-connector-python==8.0.19
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.2
pandas==1.0.2
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
poyo==0.5.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.6.1
ptyprocess==0.6.0
py4j==0.10.7
Pygments==2.5.2
pylint==2.4.4
PyMySQL==0.9.3
pyrsistent==0.16.0
pyspark==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
regex==2020.2.20
requests==2.22.0
s3transfer==0.3.3
Send2Trash==1.5.0
serverlessrepo==0.1.9
six==1.14.0
SQLAlchemy==1.3.16
terminado==0.8.3
testpath==0.4.4
tomlkit==0.5.8
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
tzlocal==2.0.0
urllib3==1.25.7
virtualenv==16.7.9
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
whichcraft==0.6.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.1.0
